+++ user_guide_src/source/changelog.rst
@@ -17,6 +17,10 @@ Release Date: Not Released
    -  Changed method ``db_connect()`` to always set the connection character set (if supported by the driver) and to fail if it can't.
    -  Removed method ``db_set_charset()`` and the ability to change a connection character set at runtime.
 
+-  Helpers
+
+   -  Updated :doc:`HTML Helper <helpers/html_helper>` function :php:func:`meta()` with support for "charset" and "property" properties.
+
 Version 3.0.4
 =============